backend/app/requirements.txt
fastapi
flask
uvicorn
psycopg2-binary
requests
python-dotenv